Moved cleanup to development.
